fix(ipc): address fourth round of PR #5 review comments
Correctness:
- notification/signal_irq: cast shift to 1u64 << (irq as u32) for
  explicit type and portable behavior
- endpoint/send: return bool (true=ok, false=invalid endpoint) instead
  of silently returning void; dispatch() now propagates u64::MAX on failure
- ipc/mod.rs: syscall 2 (ipc_send) propagates send() bool result

Docs:
- ipc/mod.rs: error convention clarified per-syscall — ipc_send/ipc_reply
  return 0 on success (not a message label); ipc_recv/ipc_call/ipc_reply_recv
  return the message label on success
- message.rs, endpoint.rs, notification.rs, capability.rs: remove stale
  "Not yet wired to main.rs" comments (all four modules are integrated)
- main.rs: correct comment — client loops yield_now() after demo, does
  not exit; idle loop is not the only remaining runnable task

+++ kernel/src/ipc/notification.rs
@@ -43,7 +43,8 @@
 //!     scancode = in(0x60)
 //!     ... process key event ...
 //! ```
-// Not yet wired to main.rs — suppress dead-code until integration.
+// Wired: notifications allocated/registered in main.rs; keyboard ISR calls signal_irq(1).
+// Keep dead-code allowance for unused APIs.
 #![allow(dead_code)]
 
 use core::sync::atomic::{AtomicU64, AtomicU8, Ordering};
@@ -176,7 +177,7 @@ pub fn signal_irq(irq: u8) {
     }
     // Set the bit for this IRQ line atomically (lock-free).
     if let Some(pending) = PENDING.get(idx as usize) {
-        pending.fetch_or(1 << irq, Ordering::Release);
+        pending.fetch_or(1u64 << (irq as u32), Ordering::Release);
     }
     // Trigger a reschedule so the blocked task runs on the next tick and
     // drains the pending bits from its wait() loop.
